15:00

+++ src/main/java/com/example/demo/anl/controller/AnalysisController.java
@@ -27,32 +27,32 @@ public class AnalysisController extends AbstractController<Analysis> {
 	final AnalysisServiceImpl service;
 
 	@PostMapping("/save")
-	public ResponseEntity<Integer> save(@RequestBody Analysis t) {
+	public ResponseEntity<Long> save(@RequestBody Analysis t) {
 		return ResponseEntity.ok(service.save(t));
 	}
 
 	@DeleteMapping("/delete")
-	public ResponseEntity<Integer> delete(@RequestBody Analysis t) {
+	public ResponseEntity<Long> delete(@RequestBody Analysis t) {
 		return ResponseEntity.ok(service.delete(t));
 	}
 
 	@GetMapping("/count")
-	public ResponseEntity<Integer> count() {
+	public ResponseEntity<Long> count() {
 		return ResponseEntity.ok(service.count());
 	}
 
 	@GetMapping("/one/{id}")
-	public ResponseEntity<Analysis> getOne(@PathVariable int id) {
+	public ResponseEntity<Analysis> getOne(@PathVariable long id) {
 		return ResponseEntity.ok(service.getOne(id));
 	}
 
 	@GetMapping("/find/{id}")
-	public ResponseEntity<Optional<Analysis>> findById(int id) {
+	public ResponseEntity<Optional<Analysis>> findById(long id) {
 		return ResponseEntity.ok(service.findById(id));
 	}
 
 	@GetMapping("/exists/{id}")
-	public ResponseEntity<Boolean> existsById(int id) {
+	public ResponseEntity<Boolean> existsById(long id) {
 		return ResponseEntity.ok(service.existsById(id));
 	}
