+++ src/main/java/com/example/demo/anl/domain/Analysis.java
@@ -26,7 +26,7 @@ public class Analysis {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "anal_num") private int analNum;
+	@Column(name = "anal_num") private long analNum;
 	
 	@ManyToOne
     @JoinColumn(name = "exhbn_num")
