+++ src/main/java/com/example/demo/anl/service/AnalysisServiceImpl.java
@@ -16,33 +16,33 @@ public class AnalysisServiceImpl extends AbstractService<Analysis> implements An
 	final AnalysisRepository analysisRepository;
 	
 	@Override 
-	public int save(Analysis Analysis) {
+	public long save(Analysis Analysis) {
 		return (analysisRepository.save(Analysis) != null) ? 1 : 0;
 	}
 	@Override 
-	public int delete(Analysis Analysis) {
+	public long delete(Analysis Analysis) {
 		analysisRepository.delete(Analysis); 
-		return(getOne(Analysis.getAnalNum()) == null) ? 1 : 0;
+		return (getOne(Analysis.getAnalNum()) == null) ? 1 : 0;
 	}
 	@Override 
-	public int count() {
-		return (int)analysisRepository.count();
+	public long count() {
+		return analysisRepository.count();
 	}
 	@Override 
-    public Analysis getOne(int id) {
+    public Analysis getOne(long id) {
     	return getOne(id);
     }
 	@Override 
-    public Optional<Analysis> findById(int id) {
+    public Optional<Analysis> findById(long id) {
     	return analysisRepository.findById(id);
 	}
     @Override 
-    public boolean existsById(int id) {
+    public boolean existsById(long id) {
     	return analysisRepository.existsById(id);
     }
     @Override 
     public List<Analysis> findAll() {
     	return analysisRepository.findAll();
     }
-    
+	
 }
