+++ src/main/java/com/example/demo/exh/domain/Exhbn.java
@@ -30,7 +30,7 @@ public class Exhbn {
 	
 	@Id 
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "exhbn_num") private int exhbnNum;
+	@Column(name = "exhbn_num") private long exhbnNum;
     @Column(name = "exhbn_title") private String exhbnTitle; 
     @Column(name = "start_date") private String startDate; 
     @Column(name = "end_date") private String endDate;
