+++ src/main/java/com/example/demo/hal/service/HallService.java
@@ -6,6 +6,6 @@ import com.example.demo.hal.domain.Hall;
 
 public interface HallService {
 	public List<Hall> findByHallNameAndHallLocation(String name, String location);
-	public int update(String hallClosed, int hallNum);
+	public long update(String hallClosed, long hallNum);
 	
 }
