+++ src/main/java/com/example/demo/rev/domain/Review.java
@@ -20,7 +20,7 @@ public class Review {
 	
 	@Id 
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "review_num") private int reviewNum;
+	@Column(name = "review_num") private long reviewNum;
 	@Column(name = "review_title") private String reviewTitle;
 	@Column(name = "review_content") private String reviewContent;
 	@Column(name = "reg_date") private String regDate;
